Enhance Me/requirements.txt
PySimpleGUI
opencv-python
numpy
Pillow
